packages/bpstd/pyproject.toml
[project]
name = "bpstd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "hhglory", email = "hhglory@outlook.com" }
]
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "django-choices>=2.0.0",
    "djangorestframework>=3.16.0",
    "eth-utils>=5.2.0",
    "geetest>=3.2.1",
    "minilog>=2.3.1",
    "pycoin>=0.92.20241201",
    "requests>=2.32.3",
    "slackclient>=2.9.4",
    "sqlalchemy>=2.0.40",
    "telegram>=0.0.1",
    "twilio>=9.5.1",
    "validators>=0.34.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
